Move eventsTableMap into its own file
Fixes a "can't access lexical declaration 'X' before initialization" error,
by resolving a kind of circular dependency.
